+++ go.mod
@@ -1,0 +1,15 @@
+module github.com/nabbat/url-shortener-server.git
+
+go 1.20
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/spf13/pflag v1.0.5
+	github.com/stretchr/testify v1.8.4
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
